requirements.txt
bottle
cherrypy
google-genai
ipykernel
jupyter_client
nbclient
nbformat
pyyaml

